requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1
bcrypt==4.0.1
email-validator==2.0.0
python-slugify==8.0.1
cryptography==41.0.7
requests==2.31.0
razorpay==1.3.0